# src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Inscripciones Académicas",
    "description": "Documentación de la API REST del MP-S2",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/estudiantes/": {
      "get": {
        "tags": [
          "Estudiantes"
        ],
        "description": "Obtiene la lista de estudiantes con filtros opcionales",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "Filtrar por email exacto",
            "type": "string"
          },
          {
            "name": "nombre",
            "in": "query",
            "description": "Filtrar por nombre exacto",
            "type": "string"
          },
          {
            "name": "bootcamp",
            "in": "query",
            "description": "Filtrar por bootcamp",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "crear un estudiante nuevo",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Datos para crear un estudiante nuevo",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "Juan Perez"
                },
                "email": {
                  "type": "string",
                  "example": "example@example.com"
                },
                "bootcamp": {
                  "type": "string",
                  "example": "Full Stack"
                }
              },
              "required": [
                "nombre",
                "email",
                "bootcamp"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/estudiantes/{id}": {
      "get": {
        "tags": [
          "Estudiantes"
        ],
        "description": "Obtiene la informacion de un estudiante en especifico por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del estudiante a buscar"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "actualizar un estudiante existente",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del estudiante a actualizar"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Datos a actualizar del estudiante",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "Juan Perez"
                },
                "email": {
                  "type": "string",
                  "example": "example@example.com"
                },
                "bootcamp": {
                  "type": "string",
                  "example": "Frontend"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "eliminar un estudiante",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del estudiante a eliminar"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}